requirements.txt
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
beautifulsoup4==4.12.3
black==24.8.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.112.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
isort==5.13.2
Jinja2==3.1.4
lxml==5.2.2
MarkupSafe==2.1.5
mypy-extensions==1.0.0
orjson==3.10.7
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
requests==2.32.3
ruff==0.5.7
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.32
starlette==0.37.2
structlog==24.4.0
trycourier==5.0.0
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.2
uvicorn==0.30.5
